Lint

+++ pkg/util/io.go
@@ -33,6 +33,7 @@ var LoadExternalFile = func(path string) ([]byte, error) {
 	return filedata, nil
 }
 
+// FormatBytes formats a number of bytes into something like "200 KiB"
 func FormatBytes(bytes uint64) string {
 	f := float64(bytes)
 	units := []string{
@@ -46,5 +47,5 @@ func FormatBytes(bytes uint64) string {
 		f /= 1024.0
 		logBase1024++
 	}
-	return fmt.Sprintf("%.3f %s", f, units[logBase1024])
+	return fmt.Sprintf("%d %s", uint64(f), units[logBase1024])
 }
